# demo-application/api/openapi.yaml
openapi: 3.0.3
info:
  title: Demo Shopping API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/auth/register:
    post:
      summary: Register (sign up)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
  /api/auth/me:
    get:
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthPrincipal'
  /api/products:
    get:
      summary: List products
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/cart:
    get:
      summary: Get current cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /api/cart/items:
    post:
      summary: Add item to cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /api/cart/items/{itemId}:
    put:
      summary: Change quantity
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeQuantityRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    delete:
      summary: Remove item
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /api/orders/checkout:
    post:
      summary: Checkout cart and place order
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders:
    get:
      summary: List orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 72
    AuthResult:
      type: object
      required: [accessToken, userId, email, role]
      properties:
        accessToken:
          type: string
        userId:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
    AuthPrincipal:
      type: object
      required: [userId, email, role]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
    Product:
      type: object
      required: [id, sku, name, priceYen]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        priceYen:
          type: integer
          format: int64
    AddCartItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 99
    ChangeQuantityRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 99
    Cart:
      type: object
      required: [cartId, userId, items, totalYen]
      properties:
        cartId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        totalYen:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
    CartItem:
      type: object
      required: [itemId, productId, unitPriceYen, quantity, subtotalYen]
      properties:
        itemId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
          nullable: true
        unitPriceYen:
          type: integer
          format: int64
        quantity:
          type: integer
        subtotalYen:
          type: integer
          format: int64
    Order:
      type: object
      required: [orderId, status, totalYen, createdAt, items]
      properties:
        orderId:
          type: string
          format: uuid
        status:
          type: string
        totalYen:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      required: [itemId, productId, productName, unitPriceYen, quantity, subtotalYen]
      properties:
        itemId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
        unitPriceYen:
          type: integer
          format: int64
        quantity:
          type: integer
        subtotalYen:
          type: integer
          format: int64